IS-1999: Use UtvidbarSykmelding on sykmelding side
Dette gir veilederne prikker på de sykmeldingene som har ekstra
informasjon.
Nå slipper de også å gå frem og tilbake mellom listen og visning av
enkeltsykmeldinger, noe som gjorde det vanskelig å huske hvilken man
hadde lest.
Bruk samme utvidbarSykmelding som i Utdrag fra sykefraværet, men flytt
den ut, siden den nå er en generell komponent.
